+++ miller-rabin.h
@@ -1,7 +1,7 @@
 #ifndef _MILLER_RABIN_H_
 #define _MILLER_RABIN_H_
 
-bool primalityTest_MillerRabin(unsigned, unsigned);
-bool witness(unsigned, unsigned);
+bool primalityTest_MillerRabin(uint64_t, uint64_t);
+bool witness(uint64_t, uint64_t);
 
 #endif
